requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0
